Fix healing button and damage application from message context menu
Relocated the helpers to @module/chat-message/helpers to be shared by the chatlog context-menu entries
